requirements.txt
streamlit
langchain
openai
streamlit-chat
python-dotenv
langchain-openai